+++ CHANGELOG.md
@@ -4,6 +4,7 @@
 
 - `ApiControllerConcern#render_json_error` improvements
 - Add Agilibox::Service object
+- Fix SMS AmazonSNS strategy bug
 
 ## 1.0.14